pyproject.toml
[project]
name = "papers2dataset"
version = "0.1.0"
description = "A tool for converting academic papers to structured datasets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "litellm>=1.55.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "curl-cffi>=0.14.0",
    "huggingface-hub>=0.26.0",
    "pandas>=2.3.3",
    "click>=8.0.0",
    "tqdm>=4.66.0",
    "pymupdf>=1.26.7",
]

[project.optional-dependencies]
dev = [
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.1",
]

[project.scripts]
papers2dataset = "papers2dataset.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
